feat: Organize utility package and add RSA asymmetric encryption/decryption
